Add Docker support and update Jenkins pipeline
Introduced a Dockerfile for containerizing the application and refactored the Jenkinsfile to use Docker for build, test, and deployment stages. Added environment variables and Docker registry integration to the pipeline. Archived the old Jenkinsfile as Jenkinsfile_old. Updated requirements.txt to include django-cors-headers, djangorestframework_simplejwt, and PyJWT for enhanced API and authentication support.